+++ drivers/sqlboiler-mysql/driver/mysql.go
@@ -386,7 +386,7 @@ func (m *MySQLDriver) TranslateColumnType(c drivers.Column) drivers.Column {
 			c.Type = "null.Float64"
 		case "boolean", "bool":
 			c.Type = "null.Bool"
-		case "date", "datetime", "timestamp", "time":
+		case "date", "datetime", "timestamp":
 			c.Type = "null.Time"
 		case "binary", "varbinary", "tinyblob", "blob", "mediumblob", "longblob":
 			c.Type = "null.Bytes"
@@ -438,7 +438,7 @@ func (m *MySQLDriver) TranslateColumnType(c drivers.Column) drivers.Column {
 			c.Type = "float64"
 		case "boolean", "bool":
 			c.Type = "bool"
-		case "date", "datetime", "timestamp", "time":
+		case "date", "datetime", "timestamp":
 			c.Type = "time.Time"
 		case "binary", "varbinary", "tinyblob", "blob", "mediumblob", "longblob":
 			c.Type = "[]byte"